backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==10.2.0
reportlab==4.0.9
python-docx==1.1.0
google-generativeai==0.3.2
python-dotenv==1.0.1
django-filter==23.5
djangorestframework-simplejwt==5.3.1 